Move locale insensitive stdtod to vespalib
