+++ vespalib/src/vespa/vespalib/locale/CMakeLists.txt
@@ -1,0 +1,7 @@
+# Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+vespa_add_library(vespalib_vespalib_locale OBJECT
+    SOURCES
+    c.cpp
+    locale.cpp
+    DEPENDS
+)
