+++ vespalib/src/vespa/vespalib/locale/c.h
@@ -1,0 +1,10 @@
+// Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+
+#pragma once
+
+namespace vespalib::locale::c {
+
+double strtod(const char *nptr, char **endptr);
+
+}
+
